Fixed Server Side Caching Issue
