+++ quell-server/__tests__/query.test.ts
@@ -2,7 +2,7 @@ import e from 'express';
 import { QuellCache } from '../src/quell';
 import { RequestType } from '../src/types';
 import schema from '../test-config/testSchema';
-import { writeToCache } from '../src/helpers/cacheHelpers';
+// import { writeToCache } from '../src/helpers/cacheHelpers';
 
 describe('server test for query', () => {
   const Quell = new QuellCache({
@@ -17,25 +17,25 @@ describe('server test for query', () => {
   beforeAll(() => {
     const promise1 = new Promise((resolve, reject) => {
       resolve(
-        writeToCache('country--1', {
+        Quell.writeToCache('country--1', {
           id: '1',
           capitol: { id: '2', name: 'DC' },
-        }, 1209600)
+        })
       );
     });
     const promise2 = new Promise((resolve, reject) => {
-      resolve(writeToCache('country--2', { id: '2' }, 1209600));
+      resolve(Quell.writeToCache('country--2', { id: '2' }));
     });
     const promise3 = new Promise((resolve, reject) => {
-      resolve(writeToCache('country--3', { id: '3' }, 1209600));
+      resolve(Quell.writeToCache('country--3', { id: '3' }));
     });
     const promise4 = new Promise((resolve, reject) => {
       resolve(
-        writeToCache('countries', [
+        Quell.writeToCache('countries', [
           'country--1',
           'country--2',
           'country--3',
-        ], 1209600)
+        ])
       );
     });
     return Promise.all([promise1, promise2, promise3, promise4]);
